Refactor: Refactoring out data and business logic.
