+++ backend/src/lambda/http/createTodo.ts
@@ -7,15 +7,9 @@ import {
 } from 'aws-lambda'
 
 import { CreateTodoRequest } from '../../requests/CreateTodoRequest'
-import { DynamoDB } from 'aws-sdk'
-import { TodoItem } from '../../models/TodoItem'
 import { createLogger } from '../../utils/logger'
+import { createTodo } from '../../businessLogic/todos'
 import { getUserIdFromJwt } from '../../auth/utils'
-import { v4 as uuid } from 'uuid'
-
-const docClient = new DynamoDB.DocumentClient()
-const TableName = process.env.TODOS_TABLE
-const fileUploadS3Bucket = process.env.FILE_UPLOAD_S3_BUCKET
 
 const logger = createLogger('http')
 
@@ -23,41 +17,20 @@ export const handler: APIGatewayProxyHandler = async (
   event: APIGatewayProxyEvent
 ): Promise<APIGatewayProxyResult> => {
   try {
-    const todoId = uuid()
     const userId = getUserIdFromJwt(event)
-    const createdAt = new Date().toJSON()
     const { name, dueDate }: CreateTodoRequest = JSON.parse(event.body)
-    const attachmentUrl = `https://${fileUploadS3Bucket}.s3.us-east-1.amazonaws.com/${todoId}`
+    logger.info('Creating todo', { name, dueDate, userId })
 
-    const Item: TodoItem = {
-      todoId,
-      userId,
-      createdAt,
-      name,
-      dueDate,
-      done: false,
-      attachmentUrl
-    }
+    const newTodo = await createTodo(name, dueDate, userId)
 
-    logger.info('Creating todo', { Item })
-
-    const todoToCreate = {
-      TableName,
-      Item
-    }
-
-    logger.info('Creating', { todoToCreate })
-
-    await docClient.put(todoToCreate).promise()
-
-    logger.info('Successfully created', { todoId })
+    logger.info('Successfully created', { newTodo })
 
     return {
       statusCode: 201,
       headers: {
         'Access-Control-Allow-Origin': '*'
       },
-      body: JSON.stringify(Item)
+      body: JSON.stringify({ item: newTodo })
     }
   } catch (e) {
     logger.info('Error', { error: e })
